Register page
This is the registration page for the blog. Working on how to send user credentials to the DB and later fetch them while logging in to the system.
Working on error handling to ensure credentials match with those in the DB and also several scenarios such as registering the same user twice.